IRON

Struct static_file::Static

pub struct Static {
    // some fields omitted
}

The static file-serving Handler.

This handler serves files from a single filesystem path, which may be absolute or relative. Incoming requests are mapped onto the filesystem by appending their URL path to the handler's root path. If the filesystem path corresponds to a regular file, the handler will attempt to serve it. Otherwise, if the path corresponds to a directory containing an index.html, the handler will attempt to serve that instead. If the path doesn't match any real object in the filesystem, the handler will return Err(Box<NoFile>). If an IO error occurs whilst attempting to serve a file, Err(Box<FileError(IoError)>) will be returned.

Methods

impl Static

fn new(root_path: Path) -> Static

Create a new instance of Static with a given root path.

If Path::new("") is given, files will be served from the current directory.