IRON

Struct static_file::StaticWithCache

pub struct StaticWithCache {
    // some fields omitted
}

Extends the Static handler with 304 caching.

If the client has a cached version of the requested file and the file hasn't been modified since it was cached, this handler returns the "304 Not Modified" response instead of the actual file.

Methods

impl StaticWithCache

fn new(root_path: Path) -> StaticWithCache

Create a new instance of StaticWithCache with a given root path.

If Path::new("") is given, files will be served from the current directory.